I made polished icons of all sizes for android!
Please allow me to upload them to the original project! I first learned about this from xda-developers' and I tried the app, and eventually I decided to male icons for it since it didn't have any, and it took me around 4.30 hours overall (not continuously) to get around it since I was trying to 1: make the images on my phone/Android and 2: find that amazing app that allowed to edit an apk and eventually compile it that was not in the Play Store anymore (not with the compile feature). And anyway the whole image-making process was very clunky and stupid as the Play Store has no good enough image-editors for such things so in the last hours of this endeavour I finally went to the only working method that was Photoshop on PC and voila I got the perfect symmetrical and clear image with transparency for Super Mario 64! Then I just went to that nice website that generated all sizes for icons of the image and just named each one of them. 

It's done, so now if placing them in res/values in your project makes it work then it'll be the way to go.